requirements.txt
altair
annotated-types
anthropic
anyio
appnope
asttokens
attrs
blinker
cachetools
certifi
charset-normalizer
click
comm
debugpy
decorator
distro
dnspython
duckduckgo-search
exceptiongroup
executing
filelock
fsspec
gitdb
gitpython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
groq
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
huggingface-hub
idna
ipykernel
ipython
jedi
jinja2
jiter
jsonschema
jsonschema-specifications
jupyter-client
jupyter-core
lxml
markdown-it-py
markupsafe
matplotlib-inline
mdurl
mpmath
narwhals
nest-asyncio
networkx
numpy
ollama
openai
packaging
pandas
parso
pexpect
pillow
platformdirs
primp
prompt-toolkit
proto-plus
protobuf
psutil
ptyprocess
pure-eval
pyarrow
pyasn1
pyasn1-modules
pydantic
pydantic-core
pydeck
pygments
pymongo
pymupdf
pyparsing
pyreqwest-impersonate
python-dateutil
python-dotenv
pytz
pyyaml
pyzmq
referencing
regex
requests
rich
rpds-py
rsa
safetensors
setuptools
six
smmap
sniffio
stack-data
streamlit
sympy
tavily-python
tenacity
tiktoken
tokenizers
toml
toolz
torch
tornado
tqdm
traitlets
transformers
typing-extensions
tzdata
uritemplate
urllib3
wcwidth
